Fix issues with internal authentication.

+++ src/main/java/com/alfredamos/listofblooddonorspringbootbackend/services/BloodStatService.java
@@ -10,6 +10,7 @@ import com.alfredamos.listofblooddonorspringbootbackend.mapper.BloodStatMapper;
 import com.alfredamos.listofblooddonorspringbootbackend.repositories.BloodStatRepository;
 import com.alfredamos.listofblooddonorspringbootbackend.repositories.UserRepository;
 import com.alfredamos.listofblooddonorspringbootbackend.utils.ResponseMessage;
+import com.alfredamos.listofblooddonorspringbootbackend.utils.SameUserAndAdmin;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -24,10 +25,14 @@ public class BloodStatService {
     private final AuthService authService;
     private final BloodStatMapper bloodStatMapper;
     private final UserRepository userRepository;
+    private final SameUserAndAdmin sameUserAndAdmin;
 
     public BloodStatDto createBloodStat(BloodStatCreate bloodStatRequest) {
         //----> Get the user associated with the creation of this blood-stat.
-        var user = authService.getUserFromContext();
+        var user = fetchAuthUser();
+
+        //----> Check for ownership or admin privilege.
+        sameUserAndAdmin.checkForOwnerShipOrAdmin(user.getId());
 
         //----> Map blood-create to blood-stat.
         var bloodStat = bloodStatMapper.toEntity(bloodStatRequest);
@@ -44,8 +49,14 @@ public class BloodStatService {
     }
 
     public ResponseMessage deleteBloodStatById(UUID id) {
-        //----> Check for existence of user.
+        //----> Check for existence of bloodStat.
         var bloodStat = fetchBloodStatById(id);
+
+        //----> Get the current-user.
+        var user = fetchAuthUser();
+
+        //----> Check for ownership or admin privilege.
+        sameUserAndAdmin.checkForOwnerShipOrAdmin(user.getId());
 
         //----> Delete the blood-stat with the giving id.
         bloodStatRepository.delete(bloodStat);
@@ -62,8 +73,11 @@ public class BloodStatService {
         var bloodStat = bloodStatMapper.toEntity(bloodStatRequest);
 
         //----> fetch the user associated with the update of this blood-stat.
-        var user = authService.getUserFromContext();
+        var user = fetchAuthUser();
         bloodStat.setUser(user);
+
+        //----> Check for ownership or admin privilege.
+        sameUserAndAdmin.checkForOwnerShipOrAdmin(user.getId());
 
         //----> save the edited blood-stat in the database.
         bloodStatRepository.save(bloodStat);
@@ -76,12 +90,18 @@ public class BloodStatService {
         //----> Check for existence of blood-stat with the giving id.
         var bloodStat = fetchBloodStatById(id);
 
+        //----> Check for ownership or admin privilege.
+        sameUserAndAdmin.checkForOwnerShipOrAdmin(bloodStat.getUser().getId());
+
         //----> Send back the result.
         return bloodStatMapper.toDTO(bloodStat);
     }
 
     public List<BloodStatDto> getAllBloodStats() {
-       //----> Get all the blood-stats from the database.
+        //----> Check for Admin privilege.
+        sameUserAndAdmin.checkForAdmin();
+
+        //----> Get all the blood-stats from the database.
         var bloodStats = bloodStatRepository.findAll();
 
         //----> Send back the results.
@@ -90,6 +110,9 @@ public class BloodStatService {
 
 
     public ResponseMessage deleteAllBloodStats(){
+        //----> Check for Admin privilege.
+        sameUserAndAdmin.checkForAdmin();
+
         //----> delete all blood-stats in the database.
         bloodStatRepository.deleteAll();
 
@@ -99,6 +122,9 @@ public class BloodStatService {
 
 
     public ResponseMessage deleteBloodStatsByUserId(UUID userId) {
+        //----> Check for same-user or admin.
+        sameUserAndAdmin.checkForOwnerShipOrAdmin(userId);
+
         //----> Fetch user with the given user-id.
         var user = fetchUserById(userId);
 
@@ -111,6 +137,9 @@ public class BloodStatService {
 
 
     public List<BloodStatDto> getBloodStatsByUserId(UUID userId) {
+        //----> Check for same-user or admin.
+        sameUserAndAdmin.checkForOwnerShipOrAdmin(userId);
+
         //----> Fetch user with the given user-id.
         var user = fetchUserById(userId);
 
@@ -124,6 +153,10 @@ public class BloodStatService {
 
     private User fetchUserById(UUID id) {
         return userRepository.findById(id).orElseThrow(() -> new NotFoundException("User not found!"));
+    }
+
+    private User fetchAuthUser(){
+        return authService.getUserFromContext();
     }
 
     private BloodStat fetchBloodStatById(UUID id) {
